+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please report any bugs you encounter by [creating a GitHub issue](https://github
   * E.g. `!r Hello! {{rules}}`
   * The symbols used can be changed with the `inlineSnippetStart` and `inlineSnippetEnd` options
   * This feature can be disabled by setting `allowInlineSnippets = off` in your config
+  * By default, the bot will refuse to send a reply with an unknown inline snippet. To disable this behavior, set `errorOnUnknownInlineSnippet = off`.
 * Plugins can now also be installed from NPM modules
   * Example: `plugins[] = npm:some-plugin-package`
 * Fix occasional bug with expiring blocks where the bot would send the expiry message multiple times